Fork AcmeMouse to translate mouse movement into mousewheel movement
When holding down the right mouse button, you can move your mouse and it
will attempt to scroll in that direction. I created this because I
uncovered an old trackball (with only two buttons) and couldn't bear
losing scrolling large texts.

The trackball scroll hook conflicts with the chorded mouse hook (probably
due to them both trying to cancel the right mouse button) and this feature
is not the type that I intend to be in AcmeMouse.

Something will have to be done though if I want to be able to run both.