Revamp scrolling
What started out as adding a log scale to the scrolling so it could become
more precise at when scrolling slowly, turned into the ability to scroll
in both axis at the same time. There is also a threshold for the speed of
mouse movement before it sends the mousewheel command allowing you to stop
the scrolling quicker.